Get through most of the build
